// client/src/consts/route.ts
enum ROUTE {
  HOME = '/',
  TEACHER = '/teacher',
  TEACHER_DETAIL = '/teacher/:teacherId',
  COACH = '/coach',
  STUDENT = '/student',
  REPORT = '/report',
  CLASS = '/class',
  LOGIN = '/login',
  NOTIFICATION = '/notification',
  SETTING = '/setting',
  USER_INFO = '/user_info',
}

export default ROUTE;